Fetch catalog of available data sets.

Fetch catalog of available data sets.

sal_fetch_catalog(rss_url = "https://cityofsalinas.opendatasoft.com/api/v2/catalog/exports/rss")

Arguments

rss_url
character, RSS feed URL.
Default: "https://cityofsalinas.opendatasoft.com/api/v2/catalog/exports/rss"

Value

tibble::tibble data frame with columns:

title

character, data set title.

dataset_id

character, data set ID to use in API requests.

pub_date

POSIXct, date-time published.

description

character, data set description.

link

character, URL linking to dataset.

References

https://cityofsalinas.opendatasoft.com/pages/homepage/

Examples

library(salinasr)
suppressPackageStartupMessages(library(tidyverse))

catalog <- sal_fetch_catalog()

catalog

#> # A tibble: 100 x 5
#>    title        dataset_id   pub_date            description        link       
#>    <chr>        <chr>        <dttm>              <chr>              <chr>      
#>  1 Sendas para~ bikeways-es  2018-07-27 09:02:02 Esta capa indica ~ http://cit~
#>  2 Salinas Out~ salinas-out~ 2018-07-18 21:24:12 This dataset cont~ http://cit~
#>  3 Eventos de ~ salinas-out~ 2018-07-18 21:24:12 Este conjunto de ~ http://cit~
#>  4 Oficina de ~ bls-smu0641~ 2018-07-05 21:38:36 Este conjunto de ~ http://cit~
#>  5 Bureau of L~ bls-smu0641~ 2018-07-05 21:38:36 This dataset from~ http://cit~
#>  6 Oficina de ~ bls-smu0600~ 2018-07-05 21:28:13 Este conjunto de ~ http://cit~
#>  7 Bureau of L~ bls-smu0600~ 2018-07-05 21:28:13 This dataset from~ http://cit~
#>  8 BETA de hue~ building-fo~ 2018-07-03 18:47:00 Esta es una versi~ http://cit~
#>  9 Building Fo~ building-fo~ 2018-07-03 18:47:00 This is a beta re~ http://cit~
#> 10 Fire Distri~ firedistric~ 2018-07-03 00:04:17 The districts as ~ http://cit~
#> # ... with 90 more rows

catalog %>% glimpse()

#> Observations: 100
#> Variables: 5
#> $ title       <chr> "Sendas para bicicletas", "Salinas Outreach Events", "E...
#> $ dataset_id  <chr> "bikeways-es", "salinas-outreach-events", "salinas-outr...
#> $ pub_date    <dttm> 2018-07-27 09:02:02, 2018-07-18 21:24:12, 2018-07-18 2...
#> $ description <chr> "Esta capa indica los diferentes tipos de sendas para b...
#> $ link        <chr> "http://cityofsalinas.opendatasoft.com/explore/dataset/...

grocery_stores <- catalog %>%
  filter(grepl("grocery stores", description))

grocery_stores %>% pull(description)

#> [1] "This map depicts the half mile walkability to grocery stores in the City of Salinas, Monterey County, California. This data is presented in the Neighborhood Vibrancy-Urban Greening Plan for the City of Salinas. Areas were calculated using a half mile walk distance to the nearest grocery store. All identified grocery stores were used, however, liquor stores, farmers markets, restaurants, and community gardens were not included."

grocery_stores %>% pull(dataset_id)

#> [1] "walkability-parks"