Output catalog information in a format more conducive to reading.

This is intended to be helpful when specific search parameters are unknown or simply for exploration of available data.

sal_research_catalog(sal_catalog, n = 5L, width = getOption("width"),
  ...)

Arguments

sal_catalog
tibble::tibble, as returned by sal_fetch_catalog.

n
integer, the number of items to return.

width
integer, number of characters per line.
Default: getOption("width")

...
Additional arguments to be passed to other methods. Not currently used.

Value

None (an invisible NULL)

Examples

library(salinasr)
suppressPackageStartupMessages(library(tidyverse))

catalog <- sal_fetch_catalog()

catalog %>%
  sal_research_catalog()

#> Title: Bikeways 
#> ID:    bikeways 
#> Date Published: 2018-08-06 09:02:02 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/bikeways/ 
#> Description: 
#> This layer depicts the different types of bike paths around the city of Salinas,
#> CA, portions of South Monterey County, and the Monterey Peninsula. This data
#> is useful for projects relating to planning, community development, traffic,
#> engineering, and numerous others. The field titled 'class' describes the type
#> of bike path: Class 1: Bike paths Class 2: Bike lane or Buffered Bike Lane Class
#> 3: Bike routes, aka sharrows or shared-lane marking Class 4: Separated Bikeway
#> Class 6: Unofficial Class 7: Unpaved Th is service was created by the GIS team
#> in 2017. Its primary function is to act as a service for the online web maps
#> that can be found on the Map Gallery within the City of Salinas website.
#> --------------------------------------------------------------------------------
#> 
#> Title: Sendas para bicicletas 
#> ID:    bikeways-es 
#> Date Published: 2018-07-27 09:02:02 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/bikeways-es/ 
#> Description: 
#> Esta capa indica los diferentes tipos de sendas para bicicletas en la ciudad
#> de Salinas, CA, partes del sur del condado de Monterrey, y la península de
#> Monterrey. Este dato es útil para proyectos relacionados con planificación,
#> desarrollo comunitario, tráfico, ingeniería y muchos otros. El campo llamado
#> “clase” describe el tipo de senda para bicicletas: Clase 1: bicisendas Clase
#> 2: carril para bicicleta o carril protegido para ciclistas Clase 3: rutas para
#> bicicletas, también conocidas como señales o indicaciones de carril compartido
#> Clase 4: Ciclovía individual Clase 6: no oficial Clase 7: no pavimentada Est e
#> servicio fue creado por el equipo de sistemas de información geográfica (GIS)
#> en el 2017. Su principal función es actuar como servicio para los mapas web en
#> línea que se encuentran en la galería de mapas dentro del sitio web de la ciudad
#> de Salinas.
#> --------------------------------------------------------------------------------
#> 
#> Title: Salinas Outreach Events 
#> ID:    salinas-outreach-events 
#> Date Published: 2018-07-18 21:24:12 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/salinas-outreach-events/ 
#> Description: 
#> This dataset contains the community outreach events conducted as a part of
#> Visión Salinas.
#> --------------------------------------------------------------------------------
#> 
#> Title: Eventos de alcance de Salinas 
#> ID:    salinas-outreach-events-es 
#> Date Published: 2018-07-18 21:24:12 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/salinas-outreach-events-es/ 
#> Description: 
#> Este conjunto de datos contiene los eventos de alcance comunitario realizados
#> como parte del programa Visión Salinas.
#> --------------------------------------------------------------------------------
#> 
#> Title: Oficina de Estadísticas Laborales: datos de empleo en sectores no agrícolas de Salinas del 2000 al 2017 
#> ID:    bls-smu06415000000000001-es 
#> Date Published: 2018-07-05 21:38:36 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/bls-smu06415000000000001-es/ 
#> Description: 
#> Este conjunto de datos de empleo, horas e ingresos a nivel estatal y en áreas
#> metropolitanas contiene datos de empleo no estacional en sectores no agrícolas
#> para Salinas, desde el año 2000 hasta el año 2017. El conjunto de datos es
#> parte del programa de Estadísticas de Empleo Actual (CES) de la Oficina de
#> Estadísticas Laborales. El programa CES es una encuesta mensual realizada
#> por la Oficina de Estadísticas Laborales. Es un programa cooperativo a nivel
#> estatal y federal que proporciona cálculos de ingresos, horas y empleo para los
#> estados y las áreas metropolitanas, en función de los registros de nómina de
#> establecimientos comerciales.
#> --------------------------------------------------------------------------------
#>