Output catalog information in a format more conducive to reading.

This is intended to be helpful when specific search parameters are unknown or simply for exploration of available data.

sal_research_catalog(sal_catalog, n = 5L, width = getOption("width"),
  ...)

Arguments

sal_catalog
tibble::tibble, as returned by sal_fetch_catalog.

n
integer, the number of items to return.

width
integer, number of characters per line.
Default: getOption("width")

...
Additional arguments to be passed to other methods. Not currently used.

Value

None (an invisible NULL)

Examples

library(salinasr)
suppressPackageStartupMessages(library(tidyverse))

catalog <- sal_fetch_catalog()

catalog %>%
  sal_research_catalog()

#> Title: Sendas para bicicletas 
#> ID:    bikeways-es 
#> Date Published: 2018-07-27 09:02:02 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/bikeways-es/ 
#> Description: 
#> Esta capa indica los diferentes tipos de sendas para bicicletas en la ciudad
#> de Salinas, CA, partes del sur del condado de Monterrey, y la península de
#> Monterrey. Este dato es útil para proyectos relacionados con planificación,
#> desarrollo comunitario, tráfico, ingeniería y muchos otros. El campo llamado
#> “clase” describe el tipo de senda para bicicletas: Clase 1: bicisendas Clase
#> 2: carril para bicicleta o carril protegido para ciclistas Clase 3: rutas para
#> bicicletas, también conocidas como señales o indicaciones de carril compartido
#> Clase 4: Ciclovía individual Clase 6: no oficial Clase 7: no pavimentada Est e
#> servicio fue creado por el equipo de sistemas de información geográfica (GIS)
#> en el 2017. Su principal función es actuar como servicio para los mapas web en
#> línea que se encuentran en la galería de mapas dentro del sitio web de la ciudad
#> de Salinas.
#> --------------------------------------------------------------------------------
#> 
#> Title: Salinas Outreach Events 
#> ID:    salinas-outreach-events 
#> Date Published: 2018-07-18 21:24:12 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/salinas-outreach-events/ 
#> Description: 
#> This dataset contains the community outreach events conducted as a part of
#> Visión Salinas.
#> --------------------------------------------------------------------------------
#> 
#> Title: Eventos de alcance de Salinas 
#> ID:    salinas-outreach-events-es 
#> Date Published: 2018-07-18 21:24:12 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/salinas-outreach-events-es/ 
#> Description: 
#> Este conjunto de datos contiene los eventos de alcance comunitario realizados
#> como parte del programa Visión Salinas.
#> --------------------------------------------------------------------------------
#> 
#> Title: Oficina de Estadísticas Laborales: datos de empleo en sectores no agrícolas de Salinas del 2000 al 2017 
#> ID:    bls-smu06415000000000001-es 
#> Date Published: 2018-07-05 21:38:36 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/bls-smu06415000000000001-es/ 
#> Description: 
#> Este conjunto de datos de empleo, horas e ingresos a nivel estatal y en áreas
#> metropolitanas contiene datos de empleo no estacional en sectores no agrícolas
#> para Salinas, desde el año 2000 hasta el año 2017. El conjunto de datos es
#> parte del programa de Estadísticas de Empleo Actual (CES) de la Oficina de
#> Estadísticas Laborales. El programa CES es una encuesta mensual realizada
#> por la Oficina de Estadísticas Laborales. Es un programa cooperativo a nivel
#> estatal y federal que proporciona cálculos de ingresos, horas y empleo para los
#> estados y las áreas metropolitanas, en función de los registros de nómina de
#> establecimientos comerciales.
#> --------------------------------------------------------------------------------
#> 
#> Title: Bureau of Labor Statistics - Salinas Non-Farm Employment Data 2000-2017 
#> ID:    bls-smu06415000000000001 
#> Date Published: 2018-07-05 21:38:36 
#> Link:
#> http://cityofsalinas.opendatasoft.com/explore/dataset/bls-smu06415000000000001/ 
#> Description: 
#> This dataset from the Bureau of Labor Statistics's State and Metro Area
#> Employment, Hours, &amp; Earnings data contains non-farm, non-seasonally
#> adjusted employment data for Salinas from 2000 to 2017. This dataset is part of
#> the Bureau of Labor Statistic's Current Employment Statistics (CES) program. The
#> CES program is a monthly survey conducted by the Bureau of Labor Statistics. It
#> is a federal and state cooperative program that provides employment, hours, and
#> earnings estimates for states and metropolitan areas based on payroll records of
#> business establishments.
#> --------------------------------------------------------------------------------
#>